requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.1.0=gpu
absl-py=0.13.0=pypi_0
aiohttp=3.5.4=pypi_0
apipkg=1.5=py_0
appdirs=1.4.3=py_1
argcomplete=1.11.1=py_0
asn1crypto=1.2.0=py36_0
astor=0.7.1=py36_0
astunparse=1.6.3=pypi_0
async-timeout=3.0.1=pypi_0
asyncio=3.4.3=pypi_0
attrs=19.1.0=pypi_0
backcall=0.1.0=py_0
basemap=1.2.1=py36h468567b_2
bcrypt=3.1.7=py36hfa6e2cd_0
blas=1.0=mkl
bleach=1.5.0=pypi_0
blinker=1.4=py36_0
bzip2=1.0.8=hfa6e2cd_2
ca-certificates=2020.10.14=0
cached-property=1.5.2=pypi_0
cachetools=4.1.1=py_0
certifi=2020.6.20=py36_0
cffi=1.12.3=py36h7a1dbc1_0
cftime=1.0.4.2=py36hc8d92b1_0
chameleon=3.6.2=py_0
chardet=3.0.4=pypi_0
clang=5.0=pypi_0
click=7.0=py_0
colorama=0.4.1=py_0
colorgram-py=1.2.0=pypi_0
comtypes=1.1.7=pypi_0
cryptography=2.7=py36h7a1dbc1_0
cudatoolkit=10.1.243=h74a9793_0
cudnn=7.6.5=cuda10.1_0
curl=7.68.0=h4496350_0
cycler=0.10.0=py_2
d3dshot=0.1.3=pypi_0
decorator=4.4.1=py_0
defusedxml=0.6.0=py_0
discord=1.0.1=pypi_0
discord-py=1.2.3=pypi_0
entrypoints=0.3=py36_1000
execnet=1.7.1=py_0
flask=1.1.1=py_1
flask-httpauth=3.3.0=py_0
flatbuffers=1.12=pypi_0
freetype=2.10.0=h563cfd7_1
fs=2.4.11=py36_0
fs.sshfs=0.12.1=py36_0
fs.webdavfs=0.3.7=py36_0
fs_filepicker=0.3.7=py36_3
furl=2.1.0=py_0
future=0.18.2=py36_0
gast=0.4.0=pypi_0
geos=3.8.0=he025d50_0
google-auth=1.22.1=py_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.1.7=pypi_0
grpcio=1.39.0=pypi_0
h5py=3.1.0=pypi_0
hdf4=4.2.13=hf8e6fe8_1003
hdf5=1.10.4=h7ebc959_0
html5lib=0.9999999=pypi_0
humanfriendly=5.0=py36_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
idna=2.8=pypi_0
idna_ssl=1.1.0=py36_0
importlib_metadata=1.1.0=py36_0
intel-openmp=2019.3=203
ipykernel=5.1.3=py36h5ca1d4c_0
ipython=7.10.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=py_1
isodate=0.6.0=py_1
itsdangerous=1.1.0=py_0
jedi=0.15.1=py36_0
jinja2=2.10.3=py_0
jpeg=9c=hfa6e2cd_1001
jplephem=2.12=pyhb958827_0
json5=0.8.5=py_0
jsonschema=3.2.0=py36_0
jupyter_client=5.3.3=py36_1
jupyter_core=4.6.1=py36_0
jupyterlab=1.2.3=py_0
jupyterlab_server=1.0.6=py_0
keras=2.6.0=pypi_0
keras-applications=1.0.8=py_1
keras-gpu=2.2.4=0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.1.0=py36he980bc4_0
krb5=1.16.4=hdd46e55_0
libcurl=7.68.0=h4496350_0
libiconv=1.15=hfa6e2cd_1005
libnetcdf=4.6.2=h396784b_1001
libpng=1.6.37=h7602738_0
libprotobuf=3.7.1=h7bd577a_0
libsodium=1.0.17=h2fa13f4_0
libssh2=1.8.2=h642c060_2
libtiff=4.1.0=h21b02b4_3
libxml2=2.9.10=h9ce36c8_0
libxslt=1.1.33=heafd4d3_0
livelossplot=0.4.1=pypi_0
lxml=4.3.3=pypi_0
lz4-c=1.8.3=he025d50_1001
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.1=py36_0
markupsafe=1.1.1=py36hfa6e2cd_0
matplotlib=3.1.2=py36_1
matplotlib-base=3.1.2=py36h2981e6d_1
menuinst=1.4.16=py36_0
mistune=0.8.4=py36hfa6e2cd_1000
mkl=2019.4=245
mkl-service=2.3.0=py36hb782905_0
mkl_fft=1.0.12=py36h14836fe_0
mkl_random=1.0.2=py36h343c172_0
mock=2.0.0=py36h9086845_0
more-itertools=7.2.0=py_0
mouseinfo=0.1.2=pypi_0
mss=5.0.0=pypi_0
msys2-conda-epoch=20160418=1
multidict=4.5.2=pypi_0
nbconvert=5.6.1=py36_0
nbformat=4.4.0=py_1
netcdf4=1.5.1.2=py36h2b7c3b1_0
notebook=6.0.1=py36_0
numpy=1.19.5=pypi_0
oauthlib=3.1.0=py_0
olefile=0.46=py_0
opencv-python=4.1.0.25=pypi_0
openssl=1.1.1h=he774522_0
opt-einsum=3.3.0=pypi_0
orderedmultidict=1.0.1=py_0
owslib=0.19.1=py_0
packaging=20.1=pypi_0
pandas=0.25.1=py36ha925a31_0
pandoc=2.8.0.1=0
pandocfilters=1.4.2=py_1
paramiko=2.7.1=py36_0
parso=0.5.1=py_0
pbr=5.1.3=py_0
pickleshare=0.7.5=py36_1000
pillow=6.0.0=pypi_0
pip=19.1.1=py36_0
proj4=5.2.0=h6538335_1006
prometheus_client=0.7.1=py_0
prompt_toolkit=3.0.2=py_0
property-cached=1.6.3=py_0
protobuf=3.17.3=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pyautogui=0.9.48=pypi_0
pycparser=2.19=py36_0
pygetwindow=0.0.8=pypi_0
pygments=2.5.2=py_0
pyjwt=1.7.1=py36_0
pymsgbox=1.0.7=pypi_0
pymysql=0.9.3=py36_0
pynacl=1.3.0=py36h2fa13f4_1001
pynput=1.6.8=pypi_0
pyopenssl=19.0.0=py36_0
pyparsing=2.4.5=py_0
pyperclip=1.7.0=pypi_0
pyproj=1.9.6=py36hfa6e2cd_1002
pyqt=5.9.2=py36h6538335_4
pyreadline=2.1=py36_1
pyrect=0.1.4=pypi_0
pyrsistent=0.15.6=py36hfa6e2cd_0
pyscreeze=0.1.26=pypi_0
pyshp=2.1.0=py_0
pysocks=1.7.1=py36_0
pytesseract=0.3.2=pypi_0
python=3.6.8=h9f7ef89_7
python-dateutil=2.8.1=py_0
python-engineio=3.0.0=py_0
python-socketio=4.3.0=py_0
pytweening=1.0.3=pypi_0
pytz=2019.3=py_0
pywin32=225=py36hfa6e2cd_0
pywinpty=0.5.7=py36_0
pyyaml=5.1=py36he774522_0
pyzmq=18.1.1=py36h16f9016_0
qt=5.9.7=hc6833c9_1
requests=2.22.0=py36_1
requests-oauthlib=1.3.0=py_0
rsa=4.6=py_0
schedule=0.6.0=pypi_0
scikit-learn=0.20.3=py36h343c172_0
scipy=1.2.1=py36h29ff71c_0
send2trash=1.5.0=py_0
setuptools=41.0.1=py36_0
sgp4=2.3=pyh5ca1d4c_0
sip=4.19.8=py36h6538335_0
six=1.15.0=pypi_0
skyfield=1.17=py_2
sqlite=3.28.0=he774522_0
tensorboard=1.14.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.6.0=pypi_0
tensorflow-estimator=2.6.0=pypi_0
tensorflow-gpu=2.6.0=pypi_0
termcolor=1.1.0=py36_1
terminado=0.8.3=py36_0
testpath=0.4.4=py_0
tk=8.6.10=hfa6e2cd_0
tornado=6.0.3=py36hfa6e2cd_0
tqdm=4.40.0=py_0
traitlets=4.3.3=py36_0
typing-extensions=3.7.4=pypi_0
typing_extensions=3.7.4.3=py_0
unicodecsv=0.14.1=py_1
urllib3=1.25.7=py36_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
wcwidth=0.1.7=py_1
webdavclient2=0.1.1=py_0
webencodings=0.5.1=py_1
websockets=6.0=pypi_0
werkzeug=0.15.2=py_0
wheel=0.37.0=pypi_0
win_inet_pton=1.1.0=py36_0
wincertstore=0.2=py36h7fe50ca_0
winpty=0.4.3=4
wrapt=1.11.2=pypi_0
xlrd=1.2.0=py36_0
xz=5.2.4=h2fa13f4_1001
yaml=0.1.7=hc54c509_2
yarl=1.3.0=pypi_0
zeromq=4.3.2=h6538335_2
zipp=0.6.0=py_0
zlib=1.2.11=h62dcd97_3
zstd=1.4.4=hd8a0e53_1
